Changed apploader to use uintptr_t for uintptr_t
git-svn-id: https://gekko-gc-emu.googlecode.com/svn/trunk@26 36332193-8412-8e5e-4a1b-12d1b2851199

+++ src/core/src/boot/apploader.cpp
@@ -6,7 +6,7 @@
 #include "hle/hle.h"
 #include "hle/hle_func.h"
 
-u32 AppLoaderPrint() {
+uintptr_t AppLoaderPrint() {
 	u32	    i;
 	char	msg[1000];
